pull function related things out to a separate file
